requirements.txt
flask
flask-mysqldb
PyJWT
xmltodict